% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/best_player.R
\name{best_player}
\alias{best_player}
\title{A best_player Function}
\usage{
best_player(x)
}
\arguments{
\item{year}{which year do you want to know the best player in.}
}
\description{
This function allows you to find the player for a certain year who scored the most points.
}
\examples{
best_player(2008)
best_player_function()
}
\keyword{most}
\keyword{point}
\keyword{scored}
\keyword{the}
