% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/function1.R
\name{charlie_black_team_games}
\alias{charlie_black_team_games}
\title{A Function}
\usage{
charlie_black_team_games(year)
}
\arguments{
\item{year}{The year that you want to find}
}
\description{
This function allows you to find whether Charlie Black, an NBA player, ever played in that year, and the games he played in each team
}
\examples{
charlie_black_team_games(1950)
}
