% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cor_in_year.r
\name{cor_in_year}
\alias{cor_in_year}
\title{A NBA Function}
\usage{
cor_in_year(year, df)
}
\arguments{
\item{year:}{integer, df: tibble.}
}
\description{
This function allows you to print correlation coefficient matrix by a certain year.
}
\examples{
cor_in_year()
}
\keyword{year}
