% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cor_mat.R
\name{cor_mat}
\alias{cor_mat}
\title{An NBA Function}
\usage{
cor_mat(year)
}
\arguments{
\item{year}{An individual year of an NBA game}
}
\description{
This function allows you to get the correlation matrix of the numeric variables in the nba dataset in a specific year.
}
\examples{
cor_mat()
}
\keyword{nba}
