% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cor_matrix_of_year.R
\name{cor_matrix_of_year}
\alias{cor_matrix_of_year}
\title{Plot the correlogram of the numeric variables of the year}
\usage{
cor_matrix_of_year(year)
}
\arguments{
\item{year}{An integer, ranges from year 1950 to year 2017}
}
\value{
Correlogram and a Correlation Matrix of all the numeric variables of the year
}
\description{
This function allows you to plot the correlogram of all numeric variables of the year, it also returns to a correlation matrix of all numeric variables of the year.
}
\examples{
cor_matrix_of_year(1999)
cor_matrix_of_year(2015)
}
\keyword{correlation}
\keyword{correlogram}
\keyword{matrix}
