% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/function3.R
\name{cor_year}
\alias{cor_year}
\title{A Function}
\usage{
cor_year(year)
}
\arguments{
\item{year}{The year that you want to search}
}
\description{
This function allows you to find the correlation matrix of all numeric vector in the NBA dataset
}
\examples{
cor_year(1950)
}
