% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/corr_num.R
\name{corr_num}
\alias{corr_num}
\title{NBA Stats Correlation Matrix}
\usage{
corr_num(year)
}
\arguments{
\item{year}{Input year of interest}
}
\description{
This function allows you input a year and analyze
the correlation between various NBA stats.
}
\examples{
corr_num()
}
\keyword{nba}
