% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/corr_year.R
\name{corr_year}
\alias{corr_year}
\title{A Function for the correlation matrix in a given year}
\usage{
corr_year(x)
}
\arguments{
\item{year}{an individual year of the NBA games}
}
\description{
This function allows you to get the correlation matrix for all numeric variables in the NBA data for that year
}
\examples{
corr_year()
}
