% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/correlation.R
\name{correlation}
\alias{correlation}
\title{Correlation Matrix}
\usage{
correlation(x)
}
\arguments{
\item{x}{The year we want to determine the statistics for.}
}
\description{
This function allows us to get the correlation matrix for all numeric variables in the data for a particular year.
}
\examples{
correlation()
}
\keyword{NBA}
\keyword{correlation}
\keyword{matrix}
