% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Matrix_function.R
\name{correlation_matrix}
\alias{correlation_matrix}
\title{Correlation Matrix Function}
\usage{
correlation_matrix(year)
}
\arguments{
\item{year}{A certain year.}

\item{df}{NBA data frame.}
}
\description{
This function returns a correlation matrix with all numeric values in a certain year.
}
\examples{
correlation_matrix()
}
\keyword{Colmatrix}
