% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/correlation_numeric.R
\name{correlation_numeric}
\alias{correlation_numeric}
\title{A correlation_numeric Function}
\usage{
correlation_numeric(x)
}
\arguments{
\item{year}{which year do you want to calculate the correlation matrix.}
}
\description{
This function allows you to calculate the correlation matrix for all numeric variables in the data for a year
}
\examples{
correlation_numeric_function
}
\keyword{correlation}
