% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/df.R
\docType{data}
\name{df}
\alias{df}
\title{NBA Player Stats.}
\format{
A data frame with 24691 rows and 52 variables:
\describe{
\item{Year}{A certain NBA season}
\item{Player}{Player's name}
...
}
}
\source{
\url{https://www.kaggle.com/drgilermo/nba-players-stats/data?select=Seasons_Stats.csv}
}
\usage{
df
}
\description{
A dataset The data-set contains aggregate individual statistics for 67 NBA seasons.
}
\keyword{datasets}
