% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fun1.R
\name{fun1}
\alias{fun1}
\title{Function 1}
\usage{
fun1(year)
}
\description{
This function allows you to get the result of the highest points of a player in a specific year
}
\examples{
fun1()
}
\keyword{highest}
\keyword{points}
