% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fun3.R
\name{fun3}
\alias{fun3}
\title{Function 3}
\usage{
fun3(year)
}
\description{
This function allows you to get the correlation matrix for all numeric variables in a specific year
}
\examples{
fun3()
}
\keyword{all}
\keyword{correlation}
\keyword{for}
\keyword{matrix}
\keyword{numeric}
\keyword{variables}
