% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/function1.R
\name{function1}
\alias{function1}
\title{Function 1}
\usage{
function1(year)
}
\arguments{
\item{Defaults}{to TRUE.}
}
\description{
This function allows you to return the age of Danny Green in a given year.
}
\examples{
function1()
}
\keyword{function1}
