% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/function2.R
\name{function2}
\alias{function2}
\title{Function 2}
\usage{
function2(year)
}
\arguments{
\item{Defaults}{to TRUE.}
}
\description{
This function allows you to return the olddest player in a given year.
}
\examples{
function1()
}
\keyword{function1}
