% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/function3.R
\name{function3}
\alias{function3}
\title{Function 3}
\usage{
function3(year)
}
\arguments{
\item{Defaults}{to TRUE.}
}
\description{
This function allows you to return the corration matrix of numeric data in a given year.
}
\examples{
function1()
}
\keyword{function1}
