% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/function_1.R
\name{function_1}
\alias{function_1}
\title{function_1}
\usage{
function_1(df, year, player)
}
\arguments{
\item{df}{This is the data frame you want to use.}

\item{year}{Choose a year you are interested in.}

\item{player}{Choose a player you are interested in.}
}
\description{
Your input for this function should be an individual year. Your output should be a data point of interest in the dataset for some player.
}
