% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/function_2.R
\name{function_2}
\alias{function_2}
\title{function_2}
\usage{
function_2(df, year)
}
\arguments{
\item{df}{This is the data frame you want to use.}

\item{year}{Choose a year you are interested in.}
}
\description{
Your input for this function should be an individual year.  Your output should be the oldest player in the NBA data for that year (if multiple players have the same age, then decide how to return a single player).
}
