% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/function_3.R
\name{function_3}
\alias{function_3}
\title{function_3}
\usage{
function_3(df, year)
}
\arguments{
\item{df}{This is the data frame you want to use.}

\item{year}{Choose a year you are interested in.}
}
\description{
Your input for this function should be an individual year.  Your function should limit the data for that year to include numeric continuous variables only (not categorical data).  Your output should be the correlation matrix for all numeric variables in the data for that year.  (Hint: use cor(data) to create correlation matrix.  If you feel adventuresome, then you can return a correlogram plot from here (Links to an external site.)Links to an external site. instead).
}
