% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/func1.R
\name{highest_FTr}
\alias{highest_FTr}
\title{An NBA Function}
\usage{
highest_FTr(year)
}
\arguments{
\item{year}{Year.}
}
\description{
This function allows you to express your love of cats.
}
\examples{
highest_FTr()
}
\keyword{NBA_func}
