% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/highest_points.R
\name{highest_points}
\alias{highest_points}
\title{Highest Points}
\usage{
highest_points(yr)
}
\arguments{
\item{yr}{Year}
}
\description{
Returns the highest points recorded in the data for a given year.
}
\examples{
highest_points(1970)

}
