% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/highest_score.R
\name{highest_score}
\alias{highest_score}
\title{A highest_score Function}
\usage{
highest_score(x)
}
\arguments{
\item{Year}{abcd hahaha}
}
\description{
Your input for this function should be an individual year.  
Your output should be the player for that year who scored the most points.
}
\examples{
highest_score()
}
\keyword{highest}
\keyword{score}
