% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Kobe_age.R
\name{Kobe_age}
\alias{Kobe_age}
\title{A Function for Kobe's age}
\usage{
Kobe_age(x)
}
\arguments{
\item{year}{an individual year of the NBA games}
}
\description{
This function allows you to find Kobe Bryant's age for any year that he's playing NBA
}
\examples{
Kobe_age()
}
