% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/more_points.R
\name{more_points}
\alias{more_points}
\title{Highest scores}
\usage{
more_points(x)
}
\arguments{
\item{x}{The year we want to determine the statistics for.}
}
\description{
This function allows you to determine which NBA player scored the highest point every year.
}
\examples{
more_points()
}
\keyword{NBA}
\keyword{highest}
\keyword{scores}
