% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/most_points.R
\name{most_points}
\alias{most_points}
\title{Who scored most?}
\usage{
most_points(x)
}
\arguments{
\item{Year}{of your choice}
}
\description{
This function shows you who scored most in the year you choose.
}
\examples{
most_points()
}
\keyword{points}
