% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/most_pts.R
\name{most_pts}
\alias{most_pts}
\title{An NBA Function}
\usage{
most_pts(year)
}
\arguments{
\item{year}{An individual year of an NBA game}
}
\description{
This function allows you to get who scored the most points in a specific year.
}
\examples{
most_pts()
}
\keyword{nba}
