% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mostpoint function.R
\name{mostpoint}
\alias{mostpoint}
\title{NBA most point function}
\usage{
mostpoint(year)
}
\arguments{
\item{year}{Indicate the year that you are interested.}
}
\description{
This function allows you to find the player with most points any year from 1950 - 2017
}
\examples{
mostpoint(1998)
}
\keyword{internal}
