% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nba_contineous.R
\name{nba_contineous}
\alias{nba_contineous}
\title{A nbapackage function}
\usage{
nba_contineous(year)
}
\arguments{
\item{year}{season you chose}
}
\description{
This function allows you to get a data point of interest in the dataset for some player from a year
}
\examples{
nba_contineous()
}
\keyword{continuous}
\keyword{correlation}
\keyword{matrix}
