% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NBA_correlation.R
\name{NBA_correlation}
\alias{NBA_correlation}
\title{NBA_correlation Function}
\usage{
NBA_correlation(input_year)
}
\arguments{
\item{input_year}{Year of NBA season}
}
\description{
This function asks for an NBA season and outputs a correlation matrix for all of the numeric variables
}
\examples{
NBA_correlation()
}
\keyword{NBA}
