% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{nba_df}
\alias{nba_df}
\title{NBA Statistics}
\format{A data frame with 52 columns.}
\source{
NBA Seasons Stats from \href{https://www.kaggle.com/drgilermo/nba-players-stats/data}{kaggle}
}
\usage{
data(nba_df)
}
\description{
NBA Statistics
}
\examples{
\dontrun{
 nba_df
}
}
\keyword{datasets}
