% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nba_oldest.R
\name{nba_oldest}
\alias{nba_oldest}
\title{A nbapackage function}
\usage{
nba_oldest(year)
}
\arguments{
\item{year}{season you chose}
}
\description{
This function allows you to get the correlation matrix for all numeric variables in the data for a year
}
\examples{
nba_oldest()
}
\keyword{oldest}
\keyword{player}
