% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NBA_oldest_player.R
\name{NBA_oldest_player}
\alias{NBA_oldest_player}
\title{NBA_oldest_player Function}
\usage{
NBA_oldest_player(input_year)
}
\arguments{
\item{input_year}{Year of NBA season}
}
\description{
This function asks for an NBA season and outputs the oldest player alphabetically for that season
}
\examples{
NBA_oldest_player()

}
\keyword{NBA}
