% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NBA_player_datapoint.R
\name{NBA_player_datapoint}
\alias{NBA_player_datapoint}
\title{NBA_player_datapoint Function}
\usage{
NBA_player_datapoint(input_year)
}
\arguments{
\item{input_year}{Year of NBA season}
}
\description{
This function asks for an NBA season and outputs the number of points Michael Jordan scored in that season
}
\examples{
NBA_player_datapoint()

}
\keyword{NBA}
