% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nba_point.R
\name{nba_point}
\alias{nba_point}
\title{A nbapackage function}
\usage{
nba_point(year)
}
\arguments{
\item{year}{season you chose}
}
\description{
This function allows you to get a data point of interest in the dataset for some player from a year
}
\examples{
nba_point()
}
\keyword{points}
