% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oldest.R
\name{oldest}
\alias{oldest}
\title{An NBA Function}
\usage{
oldest(year)
}
\arguments{
\item{year}{An individual year of an NBA game}
}
\description{
This function allows you to get the oldest player in a specific year.
}
\examples{
oldest()
}
\keyword{nba}
