% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oldest_NBAplayer_of_year.R
\name{oldest_NBAplayer_of_year}
\alias{oldest_NBAplayer_of_year}
\title{Get the oldest Player's name of the year}
\usage{
oldest_NBAplayer_of_year(year)
}
\arguments{
\item{year}{An integer, ranges from year 1950 to year 2017}
}
\value{
Name of the player, representing the oldest player of the year.
}
\description{
This function allows you to check who was the oldest Player of the year
}
\examples{
oldest_player_of_year(1999)
oldest_player_of_year(2015)
}
\keyword{oldest}
