% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oldest_player.R
\name{oldest_player}
\alias{oldest_player}
\title{Who's the oldest?}
\usage{
oldest_player(x)
}
\arguments{
\item{Year}{of your choice}
}
\description{
This function show you who's the oldest player of the year
}
\examples{
oldest_player()
}
\keyword{age}
