% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oldest_player_in_year.r
\name{oldest_player_in_year}
\alias{oldest_player_in_year}
\title{A NBA Function}
\usage{
oldest_player_in_year(year, df)
}
\arguments{
\item{year:}{integer, df: tibble.}
}
\description{
This function allows you to query an oldest player by a certain year.
}
\examples{
oldest_player_in_year()
}
\keyword{year}
