% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/function2.R
\name{oldest_year}
\alias{oldest_year}
\title{A Function}
\usage{
oldest_year(year)
}
\arguments{
\item{year}{The year that you want to search}
}
\description{
This function allows you to find the oldest player in NBA in that year
}
\examples{
oldest_year(1950)
}
