% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/player_in_year.r
\name{player_in_year}
\alias{player_in_year}
\title{A NBA Function}
\usage{
player_in_year(year, df)
}
\arguments{
\item{year:}{integer, df: tibble.}
}
\description{
This function allows you to query a data point by a certain year.
}
\examples{
player_in_year()
}
\keyword{year}
