% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_correlogram.R
\name{plot_correlogram}
\alias{plot_correlogram}
\title{Plot Correlogram Function}
\usage{
plot_correlogram(yr)
}
\arguments{
\item{yr}{Year}
}
\description{
Plots the correlogram of all numeric continuous variables in the data for a given year.
}
\examples{
plot_correlogram(1970)

}
\keyword{correlogram}
