% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PTS_top10_NBAplayers_of_year.R
\name{PTS_top10_NBAplayers_of_year}
\alias{PTS_top10_NBAplayers_of_year}
\title{Top 10 players, whit the highest PTS, of Year}
\usage{
PTS_top10_NBAplayers_of_year(year)
}
\arguments{
\item{year}{An integer, ranges from year 1950 to year 2017}
}
\value{
Names of the Top 10 players, who get the most PTS of the year.
}
\description{
This function allows you to check who are the top 10 players earning the most PTS in certain year
}
\examples{
PTS_top10_NBAplayers_of_year(1999)
PTS_top10_NBAplayers_of_year(2015)
}
\keyword{10}
\keyword{NBA}
\keyword{PTS}
\keyword{Top}
\keyword{players'}
