% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/year_best_point.R
\name{year_best_point}
\alias{year_best_point}
\title{Best Point of the Year Function}
\usage{
year_best_point(x)
}
\description{
This function calculates the best points that an NBA player could get in the given year.
}
\examples{
year_best_point(1950)
}
\keyword{points}
