% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/year_cor_matrix.R
\name{year_cor_matrix}
\alias{year_cor_matrix}
\title{Correlation Matrix of the Year Function}
\usage{
year_cor_matrix(x)
}
\description{
This function returns the correlation matrix of all numeric variables in the given year.
}
\examples{
year_cor_matrix(1950)
}
\keyword{correlation}
\keyword{matrix}
\keyword{numeric,}
