% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/year_oldest_player.R
\name{year_oldest_player}
\alias{year_oldest_player}
\title{Oldest Player of the Year Function}
\usage{
year_oldest_player(x)
}
\description{
This function returns the oldest NBA player in the given year. If multiple players have the same age, the top name in alphabetical order will be returned.
}
\examples{
year_oldest_player(1950)
}
\keyword{old,}
\keyword{player}
