% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/year_oldestplayer.R
\name{year_oldestplayer}
\alias{year_oldestplayer}
\title{Oldest NBA Player}
\usage{
year_oldestplayer(year)
}
\arguments{
\item{year}{Input year of interest}
}
\description{
This function allows you input a year and see which
NBA player was oldest in that year.
}
\examples{
year_oldestplayer()
}
\keyword{nba}
