% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/year_player.R
\name{year_player}
\alias{year_player}
\title{NBA Player with Most Scored Points}
\usage{
year_player(year)
}
\arguments{
\item{year}{Input year of interest}
}
\description{
This function allows you input a year and see which
NBA player scored the most points for that year.
}
\examples{
year_player()
}
\keyword{nba}
