% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/yr_cormatrix.R
\name{yr_cormatrix}
\alias{yr_cormatrix}
\title{Year and Correlation Matrix Function}
\usage{
yr_cormatrix(year)
}
\arguments{
\item{year}{What year are we observing? Need to define.}
}
\description{
This function creates a correlation matrix for all numeric continuous variables.
}
\examples{
yr_cormatrix()
}
\keyword{correlation}
\keyword{numeric}
\keyword{year}
