% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/yr_oldestplayer.R
\name{yr_oldestplayer}
\alias{yr_oldestplayer}
\title{Year and Oldest Player Function}
\usage{
yr_oldestplayer(year)
}
\arguments{
\item{year}{What year are we observing? Need to define.}
}
\description{
This function selects the oldest player in a given year.
If there are multiple players with the same age, the function will return the top alphabetical name.
}
\examples{
yr_oldestplayer()
}
\keyword{age}
\keyword{player}
\keyword{year}
