% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/yr_playergame.R
\name{yr_playergame}
\alias{yr_playergame}
\title{Year and Player-Game Function}
\usage{
yr_playergame(year)
}
\arguments{
\item{year}{What year are we observing? Need to define.}
}
\description{
This function selects the player with the most number of games in a given year.
}
\examples{
yr_playergame()
}
\keyword{game}
\keyword{player}
\keyword{year}
