% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cor_mat.R
\name{cor_mat}
\alias{cor_mat}
\title{Correlation Matrix}
\usage{
cor_mat(x)
}
\arguments{
\item{Year}{of your choice}
}
\description{
This function shows a correlation matrix of the year you choose.
}
\examples{
cor_mat()
}
\keyword{numeric}
