% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cor_mat.R
\name{cor_mat}
\alias{cor_mat}
\title{A Cat Function}
\usage{
cor_mat(n)
}
\description{
Your input for this function should be an individual year.  
 Your function should limit the data for that year to include 
 numeric continuous variables only (not categorical data).  
 Your output should be the correlation matrix for all numeric 
 variables in the data for that year.
}
\examples{
cor_mat()
}
\keyword{correlation}
\keyword{matrix}
