% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/correlation function.R
\name{correlation}
\alias{correlation}
\title{NBA correlation matrix(correlation)}
\usage{
correlation(year)
}
\arguments{
\item{year}{Indicate the year that you are interested.}
}
\description{
This function allows you to find the correlation among the numeric variables of all players from any year between 1950 - 2017
}
\examples{
correlation(1998)
}
\keyword{internal}
