% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/function1.R
\name{function1}
\alias{function1}
\title{Function 1}
\arguments{
\item{x}{Name of dataset}

\item{y}{Year you are looking for}
}
\description{
A Games Function
}
\details{
This function finds the player associated with the maximum number of games for a specific year.
}
\examples{
function1(data, 2001)
}
