% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/function2.R
\name{function2}
\alias{function2}
\title{Function 2
@param x Name of dataset}
\arguments{
\item{y}{Year you are looking for}
}
\description{
An Age Function
}
\details{
This function finds the oldest player for a specific year.
}
\examples{
function2(data, 2001)
}
