% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/function3.R
\name{function3}
\alias{function3}
\title{Function 3}
\arguments{
\item{x}{Name of dataset}

\item{y}{Year you are looking for}
}
\description{
A Correlation Matrix Function
}
\details{
This function creates a correlation matrix for all the numeric variables for a specific year.
}
\examples{
function3(data, 2001)
}
