% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/function3.R
\name{function3}
\alias{function3}
\title{NBA Function3}
\usage{
function3(x)
}
\description{
This function limits the data for year x to include numeric continuous variables only. The output is the correlation matrix for all numeric variables in the data for that year.
}
