% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FTr_function.R
\name{highest_FTr}
\alias{highest_FTr}
\title{Highest FTr Function}
\usage{
highest_FTr(year)
}
\arguments{
\item{year}{A certain year.}

\item{df}{NBA data frame.}
}
\description{
This function returns the player with the highest free throw rate in a certain year.
}
\examples{
highest_FTr()
}
\keyword{FTr}
