% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NBA Season Statistics.R
\docType{data}
\name{NBA}
\alias{NBA}
\title{NBA Season Statistics.}
\format{
A data frame with 24691 rows and 52 variables:
\describe{
\item{Year}{year}
\item{Player}{player name}
...
}
}
\source{
\url{https://www.kaggle.com/drgilermo/nba-players-stats/data?select=Seasons_Stats.csv}
}
\usage{
NBA
}
\description{
A dataset containing aggregate individual statistics for 67 NBA seasons. from basic box-score attributes such as points, assists, rebounds etc., to more advanced money-ball like features such as Value Over Replacement.
}
\keyword{datasets}
