% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funct1.R, R/funct2.R, R/funct3.R
\docType{data}
\name{nba}
\alias{nba}
\alias{nba}
\alias{nba}
\title{NBA Leading Scorer Function}
\format{An object of class \code{data.frame} with 24691 rows and 52 columns.}
\usage{
nba

nba

nba
}
\arguments{
\item{year}{Inputs year for output of leading scorer}

\item{year}{Inputs year for output of oldest scorer}

\item{year}{Inputs year for output of correlation matrix}
}
\description{
This function allows you to determine the leading scorer in the NBA for any given year.

This function allows you to determine the oldest player in the NBA for any given year.

This function allows you to correlate all numeric variables in the NBA dataset for any given year.
}
\examples{
highscorer()
oldest()
cor_nba()
}
\keyword{age}
\keyword{correlation}
\keyword{highscorer}
\keyword{nba}
\keyword{oldest}
\keyword{player}
\keyword{stats}
