% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nba.R
\name{nba}
\alias{nba}
\title{Net Benefit Analysis of Prediction Models and Predictive Markers}
\usage{
nba(formula, data, xstart = 0.01, xstop = 0.99, xby = 0.01,
  ymin = -0.05, harm = NULL, graph = TRUE, intervention = FALSE,
  pred.model, interventionper = 100, smooth = FALSE, loess.span = 0.1)
}
\arguments{
\item{formula}{A formula containing the outcome and predictor variables.}

\item{data}{A data frame containing the variables in the model.}

\item{xstart}{The starting probability threshold value.}

\item{xstop}{The ending probability threshold value.}

\item{xby}{The interval between \code{xstart} and \code{xstop}.}

\item{ymin}{The minimum value of net benefit (for displaying/graphing).}

\item{harm}{A vector of length equal to the vector of predictors. Specify the value of harm into model assessment.}

\item{graph}{If set to \code{TRUE}, net benefit graphic of the prediction model will be outputted on the device graphics output.}

\item{intervention}{If set to \code{TRUE}, net reduction in interventions per 100 patients are plotted.}

\item{pred.model}{Prediction models to use. This should be a vector of models such as \code{"glm"} or \code{"svm"}.}

\item{interventionper}{Number used to calculate the number of interventions avoided by. Default is \code{100}.}

\item{smooth}{If set to \code{TRUE}, a smoothed net benefit values are calculated.}

\item{loess.span}{The value of \code{span} for the loess smoothing.}
}
\value{
If the nba() function is stored in an object, the resulting output will contain an object of type list
  containing values of the net benefit of the specified prediction models in addition to the reduction in
  interventions.
}
\description{
\code{nba} implements Vicker's et al's net benefit calculation
(based on weighing the true and false positive cases) for prediction models
and predictive markers. It can be used to compare the clinical utility
of prediction models.
}
\examples{
## standard net benefit analysis
nba(cancer ~ famhistory + age + marker, data.set, pred.mode= c("glm","rpart","svm","rf"))

## loess smooth
modified_dca(cancer ~ famhistory + age + marker, data.set, pred.mode= c("glm","rpart","svm","rf"), smooth = TRUE, loess.span = 0.5)

## plot intervention reduced
modified_dca(cancer ~ famhistory + age + marker, data.set, pred.mode= c("glm","rpart","svm","rf"), smooth = TRUE, loess.span = 0.5, intervention = TRUE)
}
\references{
Vickers, A. (2008), Decision Curve Analysis: A Novel Method for Evaluating Prediction Models Vol 26,
  Issue 6, 2006.
}
\seealso{
\code{\link[DecisionCurve]{decision_curve}}
}
\author{
Chong Kim \email{chong.kim@ucdenver.edu} and Andrew Vickers \email{vickersa@mskcc.org} based on the \href{http://journals.sagepub.com/doi/abs/10.1177/0272989x06295361}{2006 Article}
}
