% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nba.R
\docType{data}
\name{nba}
\alias{nba}
\title{NBA Players stats since 1950}
\format{
A data frame with 24691 rows and 52 variables:
\describe{
\item{Year}{year, from 1950}
\item{Player}{name of nba player}
...
}
}
\source{
\url{https://www.kaggle.com/drgilermo/nba-players-stats/data}
}
\usage{
nba
}
\description{
The dataset contains aggregate individual statistics for 67 NBA seasons.
}
\keyword{datasets}
