% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oldest function.R
\name{oldest}
\alias{oldest}
\title{NBA oldest player function (oldest)}
\usage{
oldest(year)
}
\arguments{
\item{year}{Indicate the year that you are interested.}
}
\description{
This function allows you to find the oldest player any year from 1950 - 2017
}
\examples{
oldest(1998)
}
\keyword{internal}
