% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oldest.R
\name{oldest}
\alias{oldest}
\title{A oldest Function}
\usage{
oldest(m)
}
\description{
Your input for this function should be an individual year.  
Your output should be the oldest player in the NBA data for that year.
}
\examples{
oldest()
}
\keyword{oldest}
