% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oldest.R
\name{oldest}
\alias{oldest}
\title{Oldest NBA Player}
\usage{
oldest(x)
}
\arguments{
\item{x}{The year we want to determine the statistics for.}
}
\description{
This function allows you to determine which NBA player is the oldest every year.
}
\examples{
older()
}
\keyword{NBA}
\keyword{oldest}
\keyword{player}
