% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Oldest_function.R
\name{oldest_player}
\alias{oldest_player}
\title{Oldest Player Function}
\usage{
oldest_player(year)
}
\arguments{
\item{year}{A certain year.}

\item{df}{NBA data frame.}
}
\description{
This function returns the oldest player in a certain year.
}
\examples{
oldest_player()
}
\keyword{Oldest}
