% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oldest_player.R
\name{oldest_player}
\alias{oldest_player}
\title{Oldest Player}
\usage{
oldest_player(yr)
}
\arguments{
\item{yr}{Year}
}
\description{
Returns the oldest player in the data for a given year.
}
\examples{
oldest_player(1970)

}
