% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oldest_player.R
\name{oldest_player}
\alias{oldest_player}
\title{A oldest_player Function}
\usage{
oldest_player(x)
}
\arguments{
\item{year}{which year do you want to know the oldest player in.}
}
\description{
This function allows you to find the player for a certain are the oldest.
}
\examples{
oldest_player_function()
}
\keyword{oldest}
