% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oldest_player.R
\name{oldest_player}
\alias{oldest_player}
\title{A Function for the Oldest NBA player in a given year}
\usage{
oldest_player(x)
}
\arguments{
\item{year}{an individual year of the NBA games}
}
\description{
This function allows you to find the oldest player in the NBA data for that year
}
\examples{
oldest_player()
}
